Restructured and broke down Races' layout
